+++ llvm/docs/ReleaseNotes.rst
@@ -65,6 +65,14 @@ Changes to the LLVM IR
 Changes to building LLVM
 ------------------------
 
+Changes to TableGen
+-------------------
+
+* The syntax for specifying an integer range in a range list has changed.
+  The old syntax used a hyphen in the range (e.g., ``{0-9}``). The new syntax
+  uses the "`...`" range punctuator (e.g., ``{0...9}``). The hyphen syntax
+  is deprecated. The "TableGen Language Reference" document has been updated.
+
 Changes to the ARM Backend
 --------------------------